Finish refactor of command line
